requirements.dev.txt
pytest~=5.4.1
pylama
Pillow==7.0.0
wheel
zbar-py~=1.0.4
scikit-image
